changing post api. edit mode for event form
