+++ backend/apps/group/admin.py
@@ -21,7 +21,7 @@ class GroupTypeAdmin(admin.ModelAdmin):
 
 class GroupAdmin(admin.ModelAdmin):
     search_fields = ['name', 'short_name']
-    list_display = ['name', 'short_name', 'slug', 'group_type']
+    list_display = ['name', 'short_name', 'slug', 'group_type', 'can_pin']
     list_filter = ['group_type', 'creation_year', 'public', 'private',
                    'archived']
     exclude = ['members', 'subscribers']
