+++ backend/apps/group/models.py
@@ -202,7 +202,11 @@ class Group(models.Model, SlugModel):
         help_text=_("If ticked, the group page can be seen by everyone, "
                     "including non-authenticated users. Members, events and "
                     "posts still however hidden."))
-
+    can_pin = models.BooleanField(
+        verbose_name=_("Can pin"),
+        default=False,
+        help_text=_("Whether admin members can pin a post"),
+        editable=True)
     # Profile
     summary = models.CharField(
         verbose_name=_("Summary"),
@@ -324,6 +328,21 @@ class Group(models.Model, SlugModel):
                 or self.is_member(user)
                 and self.membership_set.get(student=user.student).admin
                 or self.parent and self.parent.is_admin(user))
+
+    def can_pin_post(self, user: User) -> bool:
+        """Check if the user can pin a post.
+
+        Parameters
+        ----------
+        user : User
+            The user to check for.
+
+        Returns
+        -------
+        bool
+            True if the user has admin rights.
+        """
+        return
 
     def is_member(self, user: User) -> bool:
         """Check if a user is a member for this group.
