+++ frontend/src/components/PostCard/PostCard.scss
@@ -62,4 +62,5 @@
   position: relative;
   align-self: center;
   max-height: 500px;
+  border-radius: 10px;
 }
