+++ frontend/src/translations/en-GB.json
@@ -67,6 +67,8 @@
     "createEvent": "Create event",
     "createPost": "Create post",
     "delete": "Delete",
+    "deleteEvent": "Delete event",
+    "deleteEventConfirmation": "Are you sure you want to delete this event? \nThis action is irreversible.",
     "deletePost": "Delete post",
     "description": "Description",
     "editEvent": "Edit event",
